requirements.txt
# Core
APScheduler>=3.10,<4.0
pydantic>=2.0,<3.0
flask>=3.0,<4.0
flask-cors>=4.0,<5.0
PyJWT>=2.8,<3.0
cryptography>=42.0,<44.0

# AI providers
anthropic>=0.40,<1.0
openai>=1.50,<2.0
google-genai>=1.0,<2.0

# Utilities
feedparser>=6.0,<7.0
httpx>=0.27,<1.0
PySocks>=1.7,<2.0

# Browser automation
playwright>=1.40,<2.0

# Dev / testing
pytest>=7.0,<9.0
